🔨 Change interface for `Fixture`
The previous `Fixture` interface forced both `set_up()` and `tear_down()`
functions. Although this is common in older C++-style testing frameworks,
this is actually not particularly necessary in rust -- as setup can
be done during construction, and teardown _should_ be managed by `Drop`,
as `Drop` will execute even when panicking.

This updates the interface to now only require `Fixture::prepare()`, which
is the logical setup procedure. If or when teardown is required,
the `Drop` trait should be implemented instead.