✨ Sort input parameters by signature ordering
The input parameters to `neotest` functions don't actually force
that the ordering match the ordering of the function parameters.
However, it's important that the expansion of expressions follow the
same ordering -- otherwise arguments may not match parameter types.

This adds support for sorting the parameters so that they match the
signature specified ordering.

This is a partially-completed change, as it does not yet take into
account generics; which will be handled at a later time. Functionality
is tentatively added here, but is marked as unused code.