🔨 Rename workflow `*.yml` to `*.yaml`
The more common convention tends to be to use `*.yaml` rather than
`*.yml`. This updates the files to use the more common extension.

This also adds the `workflow_dispatch` and `workflow_call` options
to the `build.yaml` at the same time, which will enable using this
workflow as dependency to other workflows (such as future deployment),
as well as calling it manually from within Github when needed.