🧹 Fix and/or suppress doctest failures
Various doctest failures are failing for the proc-macro definitions,
primarily due to the fact that proc-macros are written in terms of the
public `neotest` crate, however the documentation exists in a vacuum
where the `neotest` crate has not yet been defined.

This marks such failures as `ignore`, at least for the time being.