🔨 Add utilities for defining blocks
This library requires a lot of repeat-use of defining modules and
functions, along with calling said function. Although this can be
manually tokenized _each type_, it's much easier to help at least
define the base structure of the outline in an easy manner.

This provides some utilities to do just that by following the examples
set by the `Bracket`/`Brace`/etc syn utilities -- so that these
definitions may _surround_ a block that will be emitted.