🔨 Overhaul implementation
The implementation of the test structure has been completely
rewritten to make it easier for expanding with `section`s.

This changes the architecture to follow a more understandable
structure that can define sub-tests recursively (which makes it easier
to nest definitions of modules/tests).

The old architecture used "dispatcher", "impl", and "main" as identifiers,
whereas this structures things into:

* test:       the actual test-code, taken almost verbatim from the user
* dispatcher: code that calls the test, injecting a fixture when desired
* executor:   the actual executable test (e.g. that will be marked with
              `#[test]`). This calls the dispatcher with whatever
              parameters were requested from the test
* suite:      the collection of all test executors

A suite will always call each test executor to ensure that the suite
reflects the state of all executors. This means that if any single
executor fails, so too will the suite.

The new structure will make it far easier to add support for sections
now, since sections can be added by simply injecting the new
subsections to the tests, and the rest of the token-streaming will
come for free.