🔨 Restructure macro-generation so doctests pass
The proc-macros were previously generating tokens directly in terms of
`::neotest::` symbols, despite the fact that this crate is not within
the dependency graph of `neotest_macros` or even `neotest_common`.

Since the `neotest` crate just re-exports types from `neotest_macros`
and `neotest_common`, it is better to write things directly in terms of
these crates -- which will enable currently failing (and ignored)
doctests to pass. This makes the required modifications in order for
this to be the case.