🔨 Update definition of `Error`
The `Error` type was previously defined as a standalone error object
that was convertible from any other object; however an easier
definition is possible that requires even less work, and works seemlessly
with the `?` operator: `Box<dyn Error>`.

This updates the definition to be `Box<dyn Error>` instead, since this
now has proper conversions from any other `Error` type, and allows for
passing the object along.