🔨 Add `Result` return type to tests
The default "rust" way of handling unit tests is to panic on failure;
however, this produces very verbose error logs that are largely
undesirable in terms of output, as it complicates discovery of the
underlying cause.

Rust has supported tests that return `Result` types for some time, which
cleans up the output to simply display `Error: <message>` for the error
type, which negates the need for a panic and doesn't generate a
mandatory stack-trace. This feature is now leveraged internally to
generated `neotest` tests, which will enable building a robust
assertions library on top of this.

In order to support both user-specified and implicit return types, if
a user specifies the return type, they are responsible for explicitly
indicating the `Ok(())` at the end of a test. Otherwise, this is
implicitly generated (making a final expression a compile-error instead).

Closes #6