🔨 Rework `subtest!` syntax
The existing `subtest!` syntax was a little foreign to work with due
to the custom `|<ident>|` syntax, which feels a little unnatural for
normal rust. This effectively formed a DSL that does not exist in
other parts of the language in any capacity, and it also posed
challenges for things like introducing attributes for subtests, and
even for IDE intellisense to provide better completion-syntax.

As an alternative, a new syntax of `subtest!(<ident>, <block>)` has been
put forward for testing in this change. This drops the custom `|<ident>|`
naming convention in favor of an explicit ident first-argument, followed
by a true `block` input.

Passing a block here will provide better context for parsing and
general intellisense, and enables more room for future-growth for
supporting things like attributes in subtests as defined in #14.

Existing tests/examples have been updated to reflect the new syntax.